listo las variables de entorno y preparando todo para el paso a produccion o despliegue
