+++ src/db.js
@@ -3,10 +3,12 @@
 // mysql2/promise se utiliza para trabajar con promesas
 import { createPool } from "mysql2/promise";
 
+import { DB_DATABASE, DB_PASSWORD, DB_PORT, DB_USER, HOST } from "./config.js";
+
 export const pool = createPool({
-  host: "localhost",
-  user: "root",
-  password: "admin123",
-  port: 3306,
-  database: "companydb",
+  host: HOST,
+  user: DB_USER,
+  password: DB_PASSWORD,
+  port: DB_PORT,
+  database: DB_DATABASE,
 });
